BellmanFord_dev: Converted pull based to push based
The Bellman ford requires to get new value from the incoming edges.
The current OM has only the outgoing edges. So an upgradation was
required which was not necessary.
This situation was overcome by changing the way of updation of values
in a way that now nodes push the values rather than pull it.